formatting, remove unused asItem
